requirements.txt
beautifulsoup4
lxml
requests
selenium
webdriver-manager
supabase
python-dotenv
